pre feature cleanup
