+++ agents/craftassist/craftassist_agent.py
@@ -35,6 +35,7 @@ if __name__ == "__main__":
 
 from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.base_util import Pos, Look, npy_to_blocks_list
+from droidlet.shared_data_struct.craftassist_shared_utils import Player, Item
 from agents.droidlet_agent import DroidletAgent
 from droidlet.memory.memory_nodes import PlayerNode
 from droidlet.perception.semantic_parsing.nsp_querier import NSPQuerier
@@ -75,8 +76,6 @@ logging.getLogger().handlers.clear()
 sentry_sdk.init()  # enabled if SENTRY_DSN set in env
 DEFAULT_BEHAVIOUR_TIMEOUT = 20
 DEFAULT_FRAME = "SPEAKER"
-Player = namedtuple("Player", "entityId, name, pos, look, mainHand")
-Item = namedtuple("Item", "id, meta")
 
 
 class CraftAssistAgent(DroidletAgent):
@@ -370,7 +369,9 @@ class CraftAssistAgent(DroidletAgent):
         chat_json = False
         try:
             chat_json = json.loads(chat)
-            chat_text = list(filter(lambda x: x["id"] == "text", chat_json["content"]))[0]["content"]
+            chat_text = list(filter(lambda x: x["id"] == "text", chat_json["content"]))[0][
+                "content"
+            ]
         except:
             chat_text = chat
 
@@ -379,7 +380,7 @@ class CraftAssistAgent(DroidletAgent):
 
         if chat_json:
             chat_json["chat_memid"] = chat_memid
-            chat_json["timestamp"] = round(datetime.timestamp(datetime.now())*1000)
+            chat_json["timestamp"] = round(datetime.timestamp(datetime.now()) * 1000)
             # Send the socket event to show this reply on dashboard
             sio.emit("showAssistantReply", chat_json)
         else:
