+++ agents/craftassist/tests/base_craftassist_test_case.py
@@ -9,17 +9,16 @@ from typing import List, Sequence, Dict
 from droidlet.memory.craftassist.mc_memory_nodes import VoxelObjectNode
 from droidlet.lowlevel.minecraft.mc_util import XYZ, Block, IDM
 from droidlet.shared_data_struct.rotation import yaw_pitch
-from droidlet.lowlevel.minecraft.pyworld.world import World, Opt, flat_ground_generator
-from droidlet.lowlevel.minecraft.pyworld.utils import (
-    Player,
-    Pos,
-    Look,
-    Item,
-    Look,
-    to_relative_pos,
-)
+from droidlet.lowlevel.minecraft.pyworld.world import World
+from droidlet.base_util import Look, Pos
+from droidlet.shared_data_struct.craftassist_shared_utils import Item, Player
+from droidlet.lowlevel.minecraft.pyworld.utils import to_relative_pos, flat_ground_generator
 
 from .fake_agent import FakeAgent, FakePlayer
+
+
+class Opt:
+    pass
 
 
 class BaseCraftassistTestCase(unittest.TestCase):
