+++ droidlet/lowlevel/minecraft/pyworld/fake_mobs.py
@@ -2,7 +2,7 @@
 Copyright (c) Facebook, Inc. and its affiliates.
 """
 import numpy as np
-from .utils import Pos, Look
+from droidlet.base_util import Pos, Look
 from droidlet.shared_data_struct.craftassist_shared_utils import MOBS_BY_ID
 
 FLAT_GROUND_DEPTH = 8
