+++ droidlet/shared_data_struct/craftassist_shared_utils.py
@@ -7,6 +7,16 @@ from droidlet.base_util import adjacent, get_bounds, manhat_dist
 from droidlet.lowlevel.minecraft.craftassist_cuberite_utils.block_data import PASSABLE_BLOCKS
 from droidlet.shared_data_structs import PriorityQueue
 
+# mainHand is the item in the player or agent's hand, that will be placed by a place block action
+# it is defined in lowlevel/minecraft/client/src/types.h as Item, and has fields id, meta
+Player = namedtuple(
+    "Player", "entityId, name, pos, look, mainHand, cagent_struct", defaults=(None,) * 6
+)
+Item = namedtuple("Item", "id, meta")
+ItemStack = namedtuple("ItemStack", "item, pos, entityId")
+Mob = namedtuple("Mob", "entityId, mobType, pos, look, cagent_struct", defaults=(None,) * 5)
+
+
 CraftAssistPerceptionData = namedtuple(
     "perception_data",
     [
